+++ docs/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 ~~~~~~~~~
 
+0.45.0 (October 2023)
+---------------------
+
+* Updated pytket version requirement to 1.21.
+* Implement crosstalk noise model for AerBackend.
+* Don't include ``SimplifyInitial`` in default passes; instead make it an option
+  to ``process_circuits()``.
+
 0.44.0 (September 2023)
 -----------------------